JobsReducers index file moved to the reducers folder, now includes reducers from the Users folder as well
